Dynamic channels: Junebug and Junebug USSD
