+++ temba/channels/types/junebug/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import unicode_literals, absolute_import
+
+from .type import JunebugType  # noqa
