+++ temba/msgs/tests.py
@@ -2155,7 +2155,7 @@ class CeleryTaskTest(TembaTest):
         self.joe = self.create_contact("Joe", "+250788383444")
 
         self.channel2 = Channel.create(
-            self.org, self.user, 'RW', Channel.TYPE_JUNEBUG_USSD, None, '1234',
+            self.org, self.user, 'RW', 'JNU', None, '1234',
             config=dict(username='junebug-user', password='junebug-pass', send_url='http://example.org/'),
             uuid='00000000-0000-0000-0000-000000001234', role=Channel.ROLE_USSD)
