+++ temba/settings_common.py
@@ -1124,6 +1124,8 @@ CHANNEL_TYPES = [
     'temba.channels.types.infobip.InfobipType',
     'temba.channels.types.jasmin.JasminType',
     'temba.channels.types.jiochat.JioChatType',
+    'temba.channels.types.junebug.JunebugType',
+    'temba.channels.types.junebug_ussd.JunebugUSSDType',
     'temba.channels.types.line.LineType',
     'temba.channels.types.telegram.TelegramType',
     'temba.channels.types.twitter.TwitterType',
