+++ templates/channels/channel_claim.haml
@@ -276,22 +276,6 @@
                 Connect your <a href="http://www.kannel.org/" target="_blank">Kannel</a> instance, we'll walk you through
                 the steps necessary to get your SMSC connection working in a few minutes.
 
-    .claim-row.junebug-channel
-      .row
-        .span3
-          %a.btn.btn-primary.btn-primary-muted.btn-claim{href:"{% url 'channels.channel_claim_junebug' %}"}
-            .btn-container
-              .claim-glyph.icon-junebug
-              %p.claim-text
-                -trans "Junebug"
-
-        .span9
-          .claim-description
-            %div
-              -blocktrans
-                Connect your <a href="https://junebug.praekelt.org/" target="_blank">Junebug</a> instance that you have
-                already set up and configured.
-
     .claim-row.twiml-channel
       .row
         .span3
