+++ templates/channels/types/junebug_ussd/claim.haml
@@ -1,0 +1,1 @@
+-extends "channels/channel_claim_authenticated.html"
